+++ src/th06_102h.exe/Gui/FUN_0041adf3.c
@@ -13,7 +13,7 @@ void __thiscall Gui::FUN_0041adf3(Gui *this)
     if (g_GameManager.demo_mode == 0) {
       AnmManager::FUN_00433590(g_AnmManager,&this->impl->vm1);
       g_AsciiManager.color = (this->impl->vm1).color.color & 0xff000000 | 0xffff40;
-      if ((int)g_GameManager.current_stage < 6) {
+      if (g_GameManager.current_stage < 6) {
         local_10.x = 168.0;
         AsciiManager::AddFormatText
                   (&g_AsciiManager,&local_10,"STAGE %d",g_GameManager.current_stage);
